the basic functionality of the interpreter is implemented
Fixed the following bugs
1) grammar error incorrect parsing ((c=100));
2) write can take more than one argument
3) error related to the inability to assign a negative value to the number
4) incorrect error recognition in the label declaration
5) unary minus is now only applicable to the int type

Added the following features:
1) operators: read, break goto for

+++ ModelLangInterpreter/Ident.cpp
@@ -40,9 +40,9 @@ bool Ident::get_assign() const {
 void Ident::put_assign() {
     assign = true;
 }
-unsigned long Ident::get_value() const {
+long long Ident::get_value() const {
     return value;
 }
-void Ident::put_value(unsigned long v) {
+void Ident::put_value(long long v) {
     value = v;
 }
